Use relative equality when checking for a split
It is possible that the splitting factor is within the line segment when
using naive < comparison, but the difference is so small the factor
actually touches one of the bounds. This would manifest in the BSP as a
line segment with miniscule length which would break collision detection
and potentially rendering.