Finish the program manager interface
For the time being, this should be good enough of an interface for managing
shaders. In the future, shader and program objects should be wrapped into a
type which stores additional information. Depending on how the interface feels
for use, it might be more comfortable if the exported functions took
prog-manager implicitly via special binding.